+++ oscar-cbls/src/main/scala/oscar/cbls/lib/search/LinearSelectors.scala
@@ -21,6 +21,7 @@
 package oscar.cbls.lib.search
 
 import scala.util.Random
+import oscar.cbls._
 
 /**Provides a set of selectors to ease the development of search engines
   * @param isRandomized can be set to false if one wants a reproductible behavior of the search engine
@@ -38,7 +39,7 @@ trait LinearSelectors{
   def setRandomized(Randomized:Boolean){
     this.Randomized = Randomized
 
-    RandomGenerator = if (Randomized) { new Random() } else { new Random(0) }
+    RandomGenerator = if (Randomized) { new Random() } else { new Random(0L) }
   }
 
   @inline
@@ -51,7 +52,7 @@ trait LinearSelectors{
     * this selector is not randomized; in case of tie breaks the first one is returned
     * @param st is optional and set to true if not specified
     */
-  def selectMaxNR2[R,S](r: Iterable[R] , s: Iterable[S], f: (R,S) => Int, st: ((R,S) => Boolean) = (_:R, _:S) => true): (R,S) = {
+  def selectMaxNR2[R,S](r: Iterable[R] , s: Iterable[S], f: (R,S) => Long, st: ((R,S) => Boolean) = (_:R, _:S) => true): (R,S) = {
     selectMaxNR[(R,S)](flattenTwoIterables(r,s), (rands:(R,S)) => f(rands._1,rands._2), (rands:(R,S)) => st(rands._1,rands._2))
   }
 
@@ -59,9 +60,9 @@ trait LinearSelectors{
     * this selector is not randomized; in case of tie breaks the first one is returned
     * @param st is optional and set to true if not specified
     */
-  def selectMaxNR[R](r: Iterable[R] , f: R => Int, st: (R => Boolean) = (_:R) => true): R = {
+  def selectMaxNR[R](r: Iterable[R] , f: R => Long, st: (R => Boolean) = (_:R) => true): R = {
     var GotOne: Boolean = false
-    var MaxSoFar = 0
+    var MaxSoFar = 0L
     var Val:R = null.asInstanceOf[R]
     for (i <- r) {
       if (st(i)) {
@@ -80,7 +81,7 @@ trait LinearSelectors{
     * this selector is not randomized; in case of tie breaks the first one is returned
     * @param st is optional and set to true if not specified
     */
-  def selectMinNR2[R,S](r: Iterable[R] , s: Iterable[S], f: (R,S) => Int, st: ((R,S) => Boolean) = (_:R, _:S) => true): (R,S) = {
+  def selectMinNR2[R,S](r: Iterable[R] , s: Iterable[S], f: (R,S) => Long, st: ((R,S) => Boolean) = (_:R, _:S) => true): (R,S) = {
     selectMaxNR2[R,S](r , s, -f(_,_), st)
   }
 
@@ -88,7 +89,7 @@ trait LinearSelectors{
     * this selector is not randomized; in case of tie breaks the first one is returned
     * @param st is optional and set to true if not specified
     */
-  def selectMinNR[R](r: Iterable[R] , f: R => Int, st: (R => Boolean) = (_:R) => true): R = {
+  def selectMinNR[R](r: Iterable[R] , f: R => Long, st: (R => Boolean) = (_:R) => true): R = {
     selectMaxNR[R](r , -f(_), st)
   }
 
@@ -97,7 +98,7 @@ trait LinearSelectors{
     * @param st is optional and set to true if not specified
     * @param stop earlier stop if a couple (and output value is found)
     */
-  def selectMin2[R,S](r: Iterable[R] , s: Iterable[S], f: (R,S) => Int, st: ((R,S) => Boolean) = (_:R, _:S) => true, stop: (R,S,Int) => Boolean = (_:R,_:S,_:Int) => false): (R,S) = {
+  def selectMin2[R,S](r: Iterable[R] , s: Iterable[S], f: (R,S) => Long, st: ((R,S) => Boolean) = (_:R, _:S) => true, stop: (R,S,Long) => Boolean = (_:R,_:S,_:Long) => false): (R,S) = {
     selectMax2[R,S](r , s, -f(_,_), st, stop)
   }
 
@@ -105,7 +106,7 @@ trait LinearSelectors{
     * this selector is randomized; in case of tie breaks the returned one is chosen randomly
     * @param st is optional and set to true if not specified
     */
-  def selectMax2[R,S](r: Iterable[R] , s: Iterable[S], f: (R,S) => Int, st: ((R,S) => Boolean) = (_:R, _:S) => true, stop: (R,S,Int) => Boolean = (_:R,_:S,_:Int) => false): (R,S) = {
+  def selectMax2[R,S](r: Iterable[R] , s: Iterable[S], f: (R,S) => Long, st: ((R,S) => Boolean) = (_:R, _:S) => true, stop: (R,S,Long) => Boolean = (_:R,_:S,_:Long) => false): (R,S) = {
     selectMax[(R,S)](flattenTwoIterables(r,s), (rands:(R,S)) => f(rands._1,rands._2), (rands:(R,S)) => st(rands._1,rands._2))
   }
 
@@ -113,12 +114,12 @@ trait LinearSelectors{
     * this selector is randomized; in case of tie breaks the returned one is chosen randomly
     * @param st is optional and set to true if not specified
     */
-  def selectMax[R](r: Iterable[R] , f: R => Int, st: (R => Boolean) = (_:R) => true, stop: (R,Int) => Boolean = (_:R,_:Int) => false): R = {
-    var MaxSoFar = Int.MinValue
+  def selectMax[R](r: Iterable[R] , f: R => Long, st: (R => Boolean) = (_:R) => true, stop: (R,Long) => Boolean = (_:R,_:Long) => false): R = {
+    var MaxSoFar = Long.MinValue
     var Val:List[R] = List.empty
     for (i <- r) {
       if (st(i)) {
-        val v:Int = f(i)
+        val v:Long = f(i)
         if(stop(i,v)){
           return i
         }
@@ -138,7 +139,7 @@ trait LinearSelectors{
   /**return a couple (r,s) that is allowed: filter(r,s) is true, and minimizing f(r,s) among the allowed couples
     * this selector is randomized; in case of tie breaks the returned one is chosen randomly
     */
-  def selectMin[R,S](r: Iterable[R] , s: Iterable[S]) (f: (R,S) => Int, filter: ((R,S) => Boolean)): (R,S) = {
+  def selectMin[R,S](r: Iterable[R] , s: Iterable[S]) (f: (R,S) => Long, filter: ((R,S) => Boolean)): (R,S) = {
     selectMin[(R,S)](flattenTwoIterables(r,s))((rands:(R,S)) => f(rands._1,rands._2), (rands:(R,S)) => filter(rands._1,rands._2))
   }
 
@@ -146,8 +147,8 @@ trait LinearSelectors{
     * this selector is randomized; in case of tie breaks the returned one is chosen randomly
     * @param st is optional and set to true if not specified
     */
-  def selectMin[R](r: Iterable[R])(f: R => Int, st: (R => Boolean) = (_:R) => true): R = {
-    var MinSoFar = Int.MaxValue
+  def selectMin[R](r: Iterable[R])(f: R => Long, st: (R => Boolean) = (_:R) => true): R = {
+    var MinSoFar = Long.MaxValue
     var Val:List[R] = List.empty
     for (i <- r) {
       if (st(i)) {
@@ -182,7 +183,7 @@ trait LinearSelectors{
     if (st == null){
       var i = RandomGenerator.nextInt(r.size)
       val it = r.iterator
-      while(i > 0){it.next(); i-=1}
+      while(i > 0L){it.next(); i-=1L}
       it.next()
     }else{
       val emptyRlist:List[R]=List.empty
@@ -194,9 +195,9 @@ trait LinearSelectors{
   }
 
   /**return an element of the range. IT is selected randomly with a uniform distribution
-    * this is performed in O(1)
+    * this is performed in O(1L)
     */
-  def selectFromRange(r: Range): Int = {
+  def selectFromRange(r: Range): Long = {
     val i = RandomGenerator.nextInt(r.size)
     r.apply(i)
   }
@@ -233,14 +234,14 @@ trait LinearSelectors{
     ifNone()
   }
 
-  /**returns a randomly chosen boolean (50%-50%)*/
-  def flip(PercentTrue:Int = 50):Boolean = RandomGenerator.nextInt(100) < PercentTrue
+  /**returns a randomly chosen boolean (50L%-50L%)*/
+  def flip(PercentTrue:Long = 50L):Boolean = RandomGenerator.nextInt(100L) < PercentTrue
 
-  /**returns a random permutation of the integers in [0; N]*/
-  def getRandomPermutation(N:Int):Iterator[Int] = {
-    val intarray:Array[Int] = new Array(N)
-    for(i <- 0 until N) intarray(i)=i
-    for(i <- 0 until N){
+  /**returns a random permutation of the integers in [0L; N]*/
+  def getRandomPermutation(N:Long):Iterator[Long] = {
+    val intarray:Array[Long] = new Array(N)
+    for(i <- 0L until N) intarray(i)=i
+    for(i <- 0L until N){
       val other = selectFromRange(0 until N)
       val old = intarray(i)
       intarray(i) = intarray(other)
